webapp/requirements.txt
flask>=0.10.1
flask_debugtoolbar
psycopg2
sqlalchemy
python-dotenv
flask-bootstrap
flask-sqlalchemy
flask-nav
flask-debug
flask-cache
Flask-SQLALchemy-Session
requests
gunicorn
grequests
sqlalchemy-datatables
